requirements.txt
llama-index-core>=0.10.58
llama-index-readers-file>=0.1.9
llama-index-vector-stores-chroma>=0.1.4
chromadb>=0.4.24
numpy>=1.24
flask>=3.0
pypdf>=4.0
